+++ FalloutStrategy/FalloutEngine/FowEntity.cpp
@@ -53,9 +53,9 @@ std::vector<iPoint> FoWEntity::GetTilesInsideRadius()const
 	finishingPos.y = startingPos.y + length;
 
 	//Creates a vector with all the tiles inside a bounding box delimited by the radius
-	for (int i = startingPos.y; i < finishingPos.y; i++)
+	for(int i = startingPos.y; i < finishingPos.y; i++)
 	{
-		for (int j = startingPos.x; j < finishingPos.x; j++)
+		for(int j = startingPos.x; j < finishingPos.x; j++)
 		{
 			ret.push_back({ j,i });
 		}
@@ -73,7 +73,7 @@ void FoWEntity::ApplyMaskToTiles(std::vector<iPoint>tilesAffected)
 	unsigned short* precMask = &App->fowManager->circleMasks[boundingBoxRadius - fow_MIN_CIRCLE_RADIUS][0];
 
 	//You have to complete the code inside this for
-	for (int i = 0; i < tilesAffected.size(); i++)
+	for(size_t i = 0; i < tilesAffected.size(); i++)
 	{
 		//You have to request the fog & shroud values of each affected tile. Hint:(You can take both with a single function call)
 		FoWDataStruct* tileValue = App->fowManager->GetFoWTileState(tilesAffected[i]);
