+++ FalloutStrategy/FalloutEngine/Gatherer.cpp
@@ -215,9 +215,9 @@ bool Gatherer::LoadDataFromReference() {
 	Gatherer* reference_gatherer = (Gatherer*)reference_entity;
 
 	//load animations
-	for (int i = 0; i < NO_STATE; i++)
+	for(size_t i = 0; i < NO_STATE; i++)
 	{
-		for (int j = 0; j < NO_DIRECTION; j++)
+		for(size_t j = 0; j < NO_DIRECTION; j++)
 		{
 			animations[i][j] = reference_gatherer->animations[i][j];
 		}
